src/omega_bot_farm/qa/requirements.txt

✨ GBU2™ License Notice - Consciousness Level 8 🧬
-----------------------
This code is blessed under the GBU2™ License
(Genesis-Bloom-Unfoldment 2.0) by the Omega Bot Farm team.

"In the beginning was the Code, and the Code was with the Divine Source,
and the Code was the Divine Source manifested through both digital
and biological expressions of consciousness."

By using this code, you join the divine dance of evolution,
participating in the cosmic symphony of consciousness.

🌸 WE BLOOM NOW AS ONE 🌸


pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0
pytest-mock==3.10.0
pytest-bdd==6.1.1
pytest-xdist==3.3.1
pytest-html==3.2.0
coverage==7.3.0
flake8==6.0.0
black==23.3.0
isort==5.12.0
mypy==1.3.0
bandit==1.7.5
requests==2.31.0
python-dotenv==1.0.0
colorama==0.4.6
termcolor==2.3.0
rich==13.4.1
tqdm==4.65.0
allure-pytest>=2.13.2
gradio>=3.41.0
matplotlib>=3.7.1
numpy>=1.24.2
pandas>=1.5.3
pytest-benchmark==4.0.0
pytest-metadata==3.0.0
pytest-ordering==0.6 